MainProjectFolder/requirements.txt
anyio==4.3.0
APScheduler==3.10.4
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
bson==0.5.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==39.0.2
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.6
django-apscheduler==0.6.2
django-ckeditor==6.7.0
django-cors-headers==4.3.0
django-filter==23.3
django-import-export==3.3.1
django-js-asset==2.1.0
django-mfa2==2.8.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.7
ecdsa==0.18.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fido2==1.0.0
freetype-py==2.3.0
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
idna==3.4
inflection==0.5.1
Jinja2==3.1.3
jsonfield==3.1.0
jsonLookup==0.9.0
langdetect==1.0.9
lxml==4.9.3
MarkupPy==1.14
MarkupSafe==2.1.5
numpy==2.0.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.2
Pillow==10.0.1
psycopg2==2.9.9
pyasn1==0.6.0
PyAudio==0.2.14
pycairo==1.25.1
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.8.0
pyotp==2.9.0
pypdf==3.17.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python-u2flib-server==5.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
rlPyCairo==0.3.0
rsa==4.9
schedule==1.2.2
scipy==1.13.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
SpeechRecognition==3.10.3
sqlparse==0.4.4
starlette==0.37.2
svglib==1.5.1
tablib==3.5.0
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
ua-parser==0.18.0
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.0.6
user-agents==2.2.0
uvicorn==0.29.0
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.13
xlrd==2.0.1
xlwt==1.3.0
